reseau/requirements.txt
requests 
python-whois